Resolve localStorage ReferenceError in Next.js application
I encountered a ReferenceError related to localStorage not being defined. This was due to the usage of my custom useLocalStorage hook. To resolve this issue, I refactored the code to be inside CartContext.

I also incorporated a check for the existence of the window object using if (typeof window !== "undefined"), this ensures that localStorage is only accessed when the application is running on the client side

By moving the functionality to the CartContext and implementing this check, I successfully resolved the ReferenceError and ensure that the application functions as expected.

Additionally, I made necessary changes inside some of the functions within CartContext to ensure proper functionality. I also added comments to make everything clear