Allow null as a valid SectionName type
Modified the SectionName type definition in the types file to allow null as a valid value. This change enables flexibility in handling cases where the section name may be null, which was necessary for scenarios such as creating a page/section that is not included in the navigation bar, like the /order page. With this modification, it's possible to turn off active sections for pages/sections not present in the navigation bar.